document_processor/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
filelock>=3.0.0
tenacity>=8.0.0
pytesseract>=0.3.10
rapidfuzz>=3.0.0
pytest>=7.4.0
